+++ v3/crates/engine/tests/relationship.rs
@@ -113,6 +113,43 @@ fn test_relationships_array_with_arguments_with_graphql_config() -> anyhow::Resu
 }
 
 #[test]
+fn test_local_relationships_model_to_model_array_with_arguments_null_inputs() -> anyhow::Result<()>
+{
+    common::test_execution_expectation_for_multiple_ndc_versions(
+        "execute/relationships/array/arguments_null_inputs",
+        &[],
+        BTreeMap::from([
+            (
+                NdcVersion::V01,
+                vec!["execute/common_metadata/postgres_connector_ndc_v01_schema.json"],
+            ),
+            (
+                NdcVersion::V02,
+                vec!["execute/common_metadata/postgres_connector_ndc_v02_schema.json"],
+            ),
+        ]),
+    )
+}
+
+#[test]
+fn test_relationships_array_with_arguments_null_inputs_with_graphql_config() -> anyhow::Result<()> {
+    common::test_execution_expectation_for_multiple_ndc_versions(
+        "execute/relationships/array/arguments_null_inputs/with_graphql_config",
+        &["execute/relationships/common_metadata/graphql_config.json"],
+        BTreeMap::from([
+            (
+                NdcVersion::V01,
+                vec!["execute/common_metadata/postgres_connector_ndc_v01_schema.json"],
+            ),
+            (
+                NdcVersion::V02,
+                vec!["execute/common_metadata/postgres_connector_ndc_v02_schema.json"],
+            ),
+        ]),
+    )
+}
+
+#[test]
 fn test_local_relationships_command_to_model() -> anyhow::Result<()> {
     common::test_execution_expectation_for_multiple_ndc_versions(
         "execute/relationships/command_to_model",